@@ -1,35 +1,52 @@
 name: tfl_api_explorer
 description: The TfL API Explorer allows you to explore the TfL API.
-version: 0.3.0
-homepage: https://github.com/tnc1997/flutter-tfl-api-explorer
-author: Thomas Clark
+
+# The following line prevents the package from being accidentally published to
+# pub.dev using `pub publish`. This is preferred for private packages.
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+
+# The following defines the version and build number for your application.
+# A version number is three numbers separated by dots, like 1.2.43
+# followed by an optional build number separated by a +.
+# Both the version and the builder number may be overridden in flutter
+# build by specifying --build-name and --build-number, respectively.
+# In Android, build-name is used as versionName while build-number used as versionCode.
+# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
+# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
+# Read more about iOS versioning at
+# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
+version: 0.3.0+300
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
+  flutter_localizations:
+    sdk: flutter
+
   tfl_api_client:
     git:
       url: https://github.com/tnc1997/dart-tfl-api-client.git
-      ref: v0.3.1
+      ref: tfl-api-client_v0.4.0
 
-  cupertino_icons: ^0.1.0
-  http: ^0.12.0
-  intl: ^0.16.0
-  package_info: ^0.4.0
-  provider: ^4.0.0
-  shared_preferences: ^0.5.0
-  url_launcher: ^5.0.0
+  cupertino_icons: ^0.1.3
+  http: ^0.12.1
+  intl: ^0.16.1
+  mdi: ^3.0.0
+  package_info: ^0.4.1
+  provider: ^4.1.3
+  shared_preferences: ^0.5.7
+  url_launcher: ^5.4.10
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  mockito: ^4.0.0
-
+  mockito: ^4.1.1
+  pedantic: ^1.9.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -44,8 +61,8 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   # assets:
-  #  - images/a_dot_burr.jpeg
-  #  - images/a_dot_ham.jpeg
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
