@@ -1,42 +1,28 @@
 import 'package:flutter/material.dart';
+import 'package:mdi/mdi.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 import 'package:tfl_api_explorer/src/specifications/place_common_name_specification.dart';
+import 'package:tfl_api_explorer/src/widgets/circular_progress_indicator_future_builder.dart';
 import 'package:tfl_api_explorer/src/widgets/nullable_text.dart';
+import 'package:tfl_api_explorer/src/widgets/place_list_tile.dart';
 
 class BikePointSearchDelegate extends SearchDelegate<Place> {
   BikePointSearchDelegate({
-    @required this.bikePoints,
+    @required this.bikePointsFuture,
   });
 
-  final List<Place> bikePoints;
+  final Future<List<Place>> bikePointsFuture;
 
   @override
   ThemeData appBarTheme(BuildContext context) {
-    final theme = Theme.of(context);
-
-    return theme.copyWith(
-      inputDecorationTheme: InputDecorationTheme(
-        hintStyle: TextStyle(
-          color: theme.primaryTextTheme.title.color,
-        ),
-      ),
-      primaryColor: theme.primaryColor,
-      primaryIconTheme: theme.primaryIconTheme,
-      primaryColorBrightness: theme.primaryColorBrightness,
-      primaryTextTheme: theme.primaryTextTheme,
-      textTheme: theme.textTheme.copyWith(
-        title: theme.textTheme.title.copyWith(
-          color: theme.primaryTextTheme.title.color,
-        ),
-      ),
-    );
+    return Theme.of(context);
   }
 
   @override
   List<Widget> buildActions(BuildContext context) {
     return <Widget>[
       IconButton(
-        icon: Icon(Icons.clear),
+        icon: Icon(Mdi.close),
         onPressed: () {
           query = '';
 
@@ -49,7 +35,7 @@ class BikePointSearchDelegate extends SearchDelegate<Place> {
   @override
   Widget buildLeading(BuildContext context) {
     return IconButton(
-      icon: Icon(Icons.arrow_back),
+      icon: Icon(Mdi.arrowLeft),
       onPressed: () {
         close(context, null);
       },
@@ -58,55 +44,58 @@ class BikePointSearchDelegate extends SearchDelegate<Place> {
 
   @override
   Widget buildResults(BuildContext context) {
-    final specification = PlaceCommonNameSpecification(
-      commonName: query,
-    );
-    final bikePoints =
-        this.bikePoints.where(specification.isSatisfiedBy).toList();
+    return CircularProgressIndicatorFutureBuilder(
+      future: bikePointsFuture,
+      builder: (context, data) {
+        final specification = PlaceCommonNameSpecification(
+          commonName: query,
+        );
 
-    return ListView.builder(
-      itemBuilder: (context, index) {
-        return ListTile(
-          title: NullableText(
-            bikePoints[index].id,
-            overflow: TextOverflow.ellipsis,
-          ),
-          subtitle: NullableText(
-            bikePoints[index].commonName,
-            overflow: TextOverflow.ellipsis,
-          ),
-          onTap: () {
-            close(context, bikePoints[index]);
+        final bikePoints = data.where(specification.isSatisfiedBy).toList();
+
+        return ListView.builder(
+          itemBuilder: (context, index) {
+            return PlaceListTile(
+              place: bikePoints[index],
+              onTap: () {
+                close(context, bikePoints[index]);
+              },
+            );
           },
+          itemCount: bikePoints.length,
         );
       },
-      itemCount: bikePoints.length,
     );
   }
 
   @override
   Widget buildSuggestions(BuildContext context) {
-    final specification = PlaceCommonNameSpecification(
-      commonName: query,
-    );
-    final bikePoints =
-        this.bikePoints.where(specification.isSatisfiedBy).toList();
+    return CircularProgressIndicatorFutureBuilder(
+      future: bikePointsFuture,
+      builder: (context, data) {
+        final specification = PlaceCommonNameSpecification(
+          commonName: query,
+        );
 
-    return ListView.builder(
-      itemBuilder: (context, index) {
-        return ListTile(
-          title: NullableText(
-            bikePoints[index].commonName,
-            overflow: TextOverflow.ellipsis,
-          ),
-          onTap: () {
-            query = bikePoints[index].commonName;
+        final bikePoints = data.where(specification.isSatisfiedBy).toList();
 
-            showResults(context);
+        return ListView.builder(
+          itemBuilder: (context, index) {
+            return ListTile(
+              title: NullableText(
+                bikePoints[index].commonName,
+                overflow: TextOverflow.ellipsis,
+              ),
+              onTap: () {
+                query = bikePoints[index].commonName;
+
+                showResults(context);
+              },
+            );
           },
+          itemCount: bikePoints.length,
         );
       },
-      itemCount: bikePoints.length,
     );
   }
 }
