@@ -1,13 +1,14 @@
 import 'package:flutter/material.dart';
+import 'package:mdi/mdi.dart';
 import 'package:provider/provider.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 import 'package:tfl_api_explorer/src/delegates/bike_point_search_delegate.dart';
 import 'package:tfl_api_explorer/src/pages/bike_points/bike_point_page.dart';
-import 'package:tfl_api_explorer/src/widgets/bike_point_list_tile.dart';
 import 'package:tfl_api_explorer/src/widgets/circular_progress_indicator_future_builder.dart';
-import 'package:tfl_api_explorer/src/widgets/drawer.dart' as drawer;
+import 'package:tfl_api_explorer/src/widgets/place_list_tile.dart';
+import 'package:tfl_api_explorer/src/widgets/tfl_api_explorer_drawer.dart';
 
-class BikePointsPage extends StatelessWidget {
+class BikePointsPage extends StatefulWidget {
   static const routeName = '/bike_points';
 
   BikePointsPage({
@@ -17,56 +18,52 @@ class BikePointsPage extends StatelessWidget {
         );
 
   @override
-  Widget build(BuildContext context) {
-    final bikePointsFuture = Provider.of<TflApi>(
-      context,
-      listen: false,
-    ).bikePoints.get();
+  _BikePointsPageState createState() => _BikePointsPageState();
+}
 
+class _BikePointsPageState extends State<BikePointsPage> {
+  Future<List<Place>> _bikePointsFuture;
+
+  @override
+  Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: Text('Bike points'),
         actions: <Widget>[
-          FutureBuilder<List<Place>>(
-            future: bikePointsFuture,
-            builder: (context, snapshot) {
-              if (snapshot.hasData) {
-                return IconButton(
-                  icon: Icon(Icons.search),
-                  onPressed: () async {
-                    final bikePoint = await showSearch(
-                      context: context,
-                      delegate: BikePointSearchDelegate(
-                        bikePoints: snapshot.data,
-                      ),
-                    );
+          IconButton(
+            icon: Icon(Mdi.magnify),
+            onPressed: () async {
+              final bikePoint = await showSearch(
+                context: context,
+                delegate: BikePointSearchDelegate(
+                  bikePointsFuture: _bikePointsFuture,
+                ),
+              );
 
-                    if (bikePoint != null) {
-                      Navigator.of(context).pushNamed(
-                        BikePointPage.routeName,
-                        arguments: bikePoint,
-                      );
-                    }
-                  },
+              if (bikePoint != null) {
+                await Navigator.of(context).pushNamed(
+                  BikePointPage.routeName,
+                  arguments: bikePoint,
                 );
               }
-
-              return IconButton(
-                icon: Icon(Icons.search),
-                onPressed: null,
-              );
             },
           ),
         ],
       ),
       body: CircularProgressIndicatorFutureBuilder<List<Place>>(
-        future: bikePointsFuture,
+        future: _bikePointsFuture,
         builder: (context, data) {
           if (data != null && data.isNotEmpty) {
             return ListView.builder(
               itemBuilder: (context, index) {
-                return BikePointListTile(
-                  bikePoint: data[index],
+                return PlaceListTile(
+                  place: data[index],
+                  onTap: () {
+                    Navigator.of(context).pushNamed(
+                      BikePointPage.routeName,
+                      arguments: data[index],
+                    );
+                  },
                 );
               },
               itemCount: data.length,
@@ -78,7 +75,14 @@ class BikePointsPage extends StatelessWidget {
           }
         },
       ),
-      drawer: drawer.Drawer(),
+      drawer: TflApiExplorerDrawer(),
     );
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    
+    _bikePointsFuture = context.read<TflApi>().bikePoints.get();
   }
 }
