@@ -4,7 +4,7 @@ import 'package:tfl_api_client/tfl_api_client.dart';
 import 'package:tfl_api_explorer/src/widgets/circular_progress_indicator_future_builder.dart';
 import 'package:tfl_api_explorer/src/widgets/line_status_list_tile.dart';
 
-class LineLineStatusesPage extends StatelessWidget {
+class LineLineStatusesPage extends StatefulWidget {
   static const routeName = '/lines/:id/line_statuses';
 
   LineLineStatusesPage({
@@ -17,18 +17,20 @@ class LineLineStatusesPage extends StatelessWidget {
   final Line line;
 
   @override
-  Widget build(BuildContext context) {
-    final lineStatusesFuture = Provider.of<TflApi>(
-      context,
-      listen: false,
-    ).lines.getLineStatuses(line.id);
+  _LineLineStatusesPageState createState() => _LineLineStatusesPageState();
+}
 
+class _LineLineStatusesPageState extends State<LineLineStatusesPage> {
+  Future<List<LineStatus>> _lineStatusesFuture;
+
+  @override
+  Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: Text('Line statuses'),
       ),
       body: CircularProgressIndicatorFutureBuilder<List<LineStatus>>(
-        future: lineStatusesFuture,
+        future: _lineStatusesFuture,
         builder: (context, data) {
           return ListView.builder(
             itemBuilder: (context, index) {
@@ -41,5 +43,13 @@ class LineLineStatusesPage extends StatelessWidget {
         },
       ),
     );
+  }
+
+  @override
+  void initState() {
+    super.initState();
+
+    _lineStatusesFuture =
+        context.read<TflApi>().lines.getLineStatuses(widget.line.id);
   }
 }
