@@ -4,147 +4,146 @@ import 'package:shared_preferences/shared_preferences.dart';
 import 'package:tfl_api_explorer/src/notifiers/authentication_change_notifier.dart';
 import 'package:tfl_api_explorer/src/pages/home_page.dart';
 import 'package:tfl_api_explorer/src/widgets/circular_progress_indicator_future_builder.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:tfl_api_explorer/src/widgets/login_form.dart';
 
-class LoginPage extends StatelessWidget {
-  static const routeName = '/login';
+class LoginPage extends StatefulWidget {
+  static const routeName = 'login';
 
-  LoginPage({
-    Key key,
-  })  : _appIdFocus = FocusNode(),
-        _appIdKey = GlobalKey<FormFieldState<String>>(),
-        _appKeyFocus = FocusNode(),
-        _appKeyKey = GlobalKey<FormFieldState<String>>(),
-        _formKey = GlobalKey<FormState>(),
-        super(
-          key: key,
-        );
+  @override
+  _LoginPageState createState() => _LoginPageState();
+}
 
-  final FocusNode _appIdFocus;
+class _LoginPageState extends State<LoginPage> {
+  TextEditingController _appIdController;
+  TextEditingController _appKeyController;
+  Future<SharedPreferences> _preferencesFuture;
 
-  final GlobalKey<FormFieldState<String>> _appIdKey;
-
-  final FocusNode _appKeyFocus;
-
-  final GlobalKey<FormFieldState<String>> _appKeyKey;
-
-  final GlobalKey<FormState> _formKey;
+  final _formKey = GlobalKey<FormState>();
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: Text('Login'),
-      ),
-      body: CircularProgressIndicatorFutureBuilder<SharedPreferences>(
-        future: SharedPreferences.getInstance(),
-        builder: (context, data) {
-          void login() {
-            if (_formKey.currentState.validate()) {
-              final appId = _appIdKey.currentState.value;
-              final appKey = _appKeyKey.currentState.value;
-
-              data.setString('APP_ID', appId);
-              data.setString('APP_KEY', appKey);
-
-              Provider.of<AuthenticationChangeNotifier>(
-                context,
-                listen: false,
-              ).login(
-                appId,
-                appKey,
-              );
-
-              Navigator.of(context).pushReplacementNamed(HomePage.routeName);
-            }
-          }
-
-          return Padding(
-            padding: const EdgeInsets.symmetric(
-              vertical: 8.0,
-              horizontal: 16.0,
-            ),
-            child: Form(
-              key: _formKey,
-              child: Column(
-                children: <Widget>[
-                  TextFormField(
-                    key: _appIdKey,
-                    initialValue: data.getString('APP_ID'),
-                    focusNode: _appIdFocus,
-                    decoration: InputDecoration(
-                      labelText: 'App id',
+      body: SafeArea(
+        child: CircularProgressIndicatorFutureBuilder<SharedPreferences>(
+          future: _preferencesFuture,
+          builder: (context, preferences) {
+            return LayoutBuilder(
+              builder: (context, constraints) {
+                if (constraints.maxWidth > 768) {
+                  return Center(
+                    child: SizedBox(
+                      width: constraints.maxWidth / 2,
+                      child: Card(
+                        child: Padding(
+                          padding: const EdgeInsets.symmetric(
+                            vertical: 8.0,
+                            horizontal: 16.0,
+                          ),
+                          child: Column(
+                            mainAxisSize: MainAxisSize.min,
+                            crossAxisAlignment: CrossAxisAlignment.stretch,
+                            children: <Widget>[
+                              Padding(
+                                padding: const EdgeInsets.only(
+                                  top: 8.0,
+                                  bottom: 16.0,
+                                ),
+                                child: Text(
+                                  'Sign in',
+                                  style: Theme.of(context).textTheme.headline6,
+                                ),
+                              ),
+                              LoginForm(
+                                formKey: _formKey,
+                                appIdController: _appIdController,
+                                appKeyController: _appKeyController,
+                                onSubmitted: _login,
+                              ),
+                            ],
+                          ),
+                        ),
+                      ),
                     ),
-                    textInputAction: TextInputAction.next,
-                    onEditingComplete: () {
-                      if (_appIdKey.currentState.validate()) {
-                        _appIdFocus.unfocus();
-
-                        FocusScope.of(context).requestFocus(_appKeyFocus);
-                      }
-                    },
-                    validator: (value) {
-                      return value.isEmpty ? 'Please enter your app ID.' : null;
-                    },
-                  ),
-                  TextFormField(
-                    key: _appKeyKey,
-                    initialValue: data.getString('APP_KEY'),
-                    focusNode: _appKeyFocus,
-                    decoration: InputDecoration(
-                      labelText: 'App key',
+                  );
+                } else {
+                  return Padding(
+                    padding: const EdgeInsets.symmetric(
+                      vertical: 8.0,
+                      horizontal: 16.0,
                     ),
-                    textInputAction: TextInputAction.go,
-                    onEditingComplete: () async {
-                      if (_appKeyKey.currentState.validate()) {
-                        _appKeyFocus.unfocus();
-
-                        login();
-                      }
-                    },
-                    validator: (value) {
-                      return value.isEmpty
-                          ? 'Please enter your app key.'
-                          : null;
-                    },
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.only(top: 8.0),
-                    child: Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      crossAxisAlignment: CrossAxisAlignment.stretch,
                       children: <Widget>[
-                        Expanded(
-                          flex: 2,
-                          child: RaisedButton(
-                            onPressed: () async {
-                              await launch(
-                                'https://api-portal.tfl.gov.uk/signup',
-                              );
-                            },
-                            child: Text('Sign up'),
+                        Padding(
+                          padding: const EdgeInsets.only(
+                            top: 8.0,
+                            bottom: 16.0,
+                          ),
+                          child: Text(
+                            'Sign in',
+                            style: Theme.of(context).textTheme.headline6,
                           ),
                         ),
-                        Spacer(
-                          flex: 1,
-                        ),
-                        Expanded(
-                          flex: 2,
-                          child: RaisedButton(
-                            onPressed: () async {
-                              login();
-                            },
-                            child: Text('Sign in'),
-                          ),
+                        LoginForm(
+                          formKey: _formKey,
+                          appIdController: _appIdController,
+                          appKeyController: _appKeyController,
+                          onSubmitted: _login,
                         ),
                       ],
                     ),
-                  )
-                ],
-              ),
-            ),
-          );
-        },
+                  );
+                }
+              },
+            );
+          },
+        ),
       ),
     );
+  }
+
+  @override
+  void dispose() {
+    _appIdController.dispose();
+    _appKeyController.dispose();
+
+    super.dispose();
+  }
+
+  @override
+  void initState() {
+    super.initState();
+
+    _appIdController = TextEditingController();
+    _appKeyController = TextEditingController();
+
+    _preferencesFuture = SharedPreferences.getInstance().then((preferences) {
+      if (preferences.containsKey('APP_ID')) {
+        _appIdController.text = preferences.getString('APP_ID');
+      }
+
+      if (preferences.containsKey('APP_KEY')) {
+        _appKeyController.text = preferences.getString('APP_KEY');
+      }
+
+      return preferences;
+    });
+  }
+
+  Future<void> _login() async {
+    if (_formKey.currentState.validate()) {
+      final preferences = await SharedPreferences.getInstance();
+
+      final appId = _appIdController.text;
+      final appKey = _appKeyController.text;
+
+      await preferences.setString('APP_ID', appId);
+      await preferences.setString('APP_KEY', appKey);
+
+      context.read<AuthenticationChangeNotifier>().login(appId, appKey);
+
+      await Navigator.of(context).pushReplacementNamed(HomePage.routeName);
+    }
   }
 }
