@@ -3,23 +3,22 @@ import 'package:package_info/package_info.dart';
 import 'package:tfl_api_explorer/src/widgets/circular_progress_indicator_future_builder.dart';
 import 'package:tfl_api_explorer/src/widgets/nullable_text.dart';
 
-class AboutPage extends StatelessWidget {
-  AboutPage({
-    Key key,
-  }) : super(
-          key: key,
-        );
+class AboutPage extends StatefulWidget {
+  @override
+  _AboutPageState createState() => _AboutPageState();
+}
+
+class _AboutPageState extends State<AboutPage> {
+  Future<PackageInfo> _packageInfoFuture;
 
   @override
   Widget build(BuildContext context) {
-    final packageInfoFuture = PackageInfo.fromPlatform();
-
     return Scaffold(
       appBar: AppBar(
         title: Text('About'),
       ),
       body: CircularProgressIndicatorFutureBuilder<PackageInfo>(
-        future: packageInfoFuture,
+        future: _packageInfoFuture,
         builder: (context, data) {
           return ListView(
             children: <Widget>[
@@ -36,5 +35,12 @@ class AboutPage extends StatelessWidget {
         },
       ),
     );
+  }
+
+  @override
+  void initState() {
+    super.initState();
+
+    _packageInfoFuture = PackageInfo.fromPlatform();
   }
 }
