@@ -29,7 +29,7 @@ class StopPointModesPage extends StatelessWidget {
             ),
           );
         },
-        itemCount: stopPoint.modes?.length,
+        itemCount: stopPoint.modes.length,
       ),
     );
   }
